added image urls parsing from css, initial version

+++ src/utils/httpUtils.ts
@@ -2,6 +2,8 @@ import * as http from 'http';
 import * as https from 'https';
 import * as fs from 'fs';
 import * as Promise from 'bluebird';
+import * as path from 'path';
+import * as pathUtils from './pathUtils';
 
 export function downloadFile(fromUrl:string, destPath:string){
     if(fs.existsSync(destPath)){
@@ -17,8 +19,11 @@ export function downloadFile(fromUrl:string, destPath:string){
                     resolve(destPath);
                 });
             }, onError = (err:any) =>{
-                fs.unlinkSync(destPath);
-                reject();
+                try{
+                    if(fs.existsSync(destPath))
+                        fs.unlinkSync(destPath);
+                }catch(err2){}
+                reject(err);
             };
 
         if(fromUrl.startsWith('https'))
@@ -26,4 +31,16 @@ export function downloadFile(fromUrl:string, destPath:string){
         else
             http.get(fromUrl, cb).on('error', onError);
     });
+}
+
+
+export function mirrorDownload(url:string, rootPath:string){
+    const srcUrl = new URL(url),
+    srcFilePath = path.dirname(srcUrl.pathname),
+    destFileFolder = path.join(rootPath, srcFilePath),
+    filename = path.basename(srcUrl.pathname),
+    destFilePath = path.join(destFileFolder, filename);
+
+    pathUtils.ensurePath(destFileFolder);
+    return downloadFile(url, destFilePath);
 }
